% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{bin_index}
\alias{bin_index}
\title{binary index}
\usage{
bin_index(x, mingap = 3)
}
\arguments{
\item{x}{integer vector, Ranks.}

\item{mingap}{minimal gap to generate binary pattern.}
}
\value{
binary index with "DE"(differential exprresion) or "UC"(unclear).
}
\description{
Function to generate binary index(0/1) of ranked data. Define gaps as the
diferrence of sorted vectors(ranks, low to high), and ranks over the maximal
gap set to 1, otherwise set to 0, always select the first maximal gap. If
specify \code{mingap}, then ranks over the \code{mingap} set to 1, othewise
set to 0. Ranks value from function:\code{psi_seq_rank}, \code{expr_seq_rank},
\code{expr_quant_rank}. Ib is the sum of binary pattern if vector has mingap,
otherwise Ib = 0.
}
