% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{bin_pattern}
\alias{bin_pattern}
\title{binary pattern}
\usage{
bin_pattern(x, mingap = 3)
}
\arguments{
\item{x}{integer vector, Ranks.}

\item{mingap}{minimal gap to generate binary pattern.}
}
\value{
binary pattern, which can be classified into "DE"(differential
exprresion, with mingap) or "UC"(unclear, ) or "HK"(house keeping, all in
the same rank).
}
\description{
Function to generate binary pattern(0/1) of ranked data. Define gaps as the
diferrence of sorted vectors(ranks, low to high), and ranks over the
maximal gap set to 1, otherwise set to 0, always select the first maximal
gap. If specify \code{mingap}, then ranks over the \code{mingap} set to 1,
othewise set to 0. Ranks value from function:\code{psi_seq_rank},
\code{expr_seq_rank}, \code{expr_quant_rank}.
}
