% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{chk_sub}
\alias{chk_sub}
\title{check whether exsiting substring and return messages.}
\usage{
chk_sub(vect)
}
\arguments{
\item{vect}{vector th access.}
}
\description{
check whether exsiting substring and return messages.
}
