% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{demo_psi}
\alias{demo_psi}
\title{A psi demo of AS events in 17 tissues}
\format{A data.frame with 5000 rows and 35variables:
\describe{
  \item{AS_events}{Alternative splicing has same start}
  \item{sample_A_1}{sample_A, replicate 1}
  \item{sample_A_2}{sample_A, replicate 2}
  ...
}}
\usage{
demo_psi
}
\description{
A dataset containing the alternative splicing events and their
percentage-of-splicing-in.
}
\keyword{datasets}
