% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{demo_tpm}
\alias{demo_tpm}
\title{A gene expression demo of TPM in 16 tissues}
\format{A data.frame with 5000 rows and 33 variables:
\describe{
  \item{gene_id}{ENSEMBL gene id}
  \item{sample_A_1}{sample_A, replicate 1}
  \item{sample_A_2}{sample_A, replicate 2}
  ...
}}
\usage{
demo_tpm
}
\description{
A dataset containing the gene names and their expression in tissues of TPM.
}
\keyword{datasets}
