% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{expr_bks_rank}
\alias{expr_bks_rank}
\title{Ranking gene expression according to given breaks}
\usage{
expr_bks_rank(df, bks)
}
\arguments{
\item{df}{data.frame.}

\item{bks}{vector, numeric. Intervals to cut according to \code{bks}.
Finally generate \code{length(bks)-1} ranks.}
}
\value{
data.frame contain ranks from \code{df}.
}
\description{
Function to cut data.frame range intointervals according to given breaks
points, maximal point and minimal point, then rank them for their value.
This function is design for raw expression value(RPKM/FPKM/TPM).
}
