% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{expr_quant_rank}
\alias{expr_quant_rank}
\title{Ranking gene expression according to equal-density-intervals(quantiles)}
\usage{
expr_quant_rank(df, n = 10, min = 0, max = 12)
}
\arguments{
\item{df}{data.frame.}

\item{n}{number of medium intervals to generated (all \code{n}+2). Default
10.}

\item{min}{psi under \code{min} will be set to 0. Default 0.}

\item{max}{interval width. Default 12, 2^12 = 4096.}
}
\value{
data.frame contain ranks from \code{df}.
}
\description{
Function to cut data.frame range into n equal-density-interval(quantiles)
points, maximal point and minimal point, then rank them for their value.
This function is design for processed data, like log2 transformed gene
expression value(RPKM/FPKM/TPM).
}
