% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_prep.R
\name{fmt_df}
\alias{fmt_df}
\title{format data.frame}
\usage{
fmt_df(df, tissues, identifier)
}
\arguments{
\item{df}{numeric.}

\item{tissues}{charactors. Colnames you want in final results.}

\item{identifier}{which col to rownames.}
}
\description{
Generate a data.frame, which only contain exprssion data of \code{tissues}
and rownames is value of \code{identifier} (gene_id, events).
}
