% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{multiplot}
\alias{multiplot}
\title{multi-plot}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{plots}

\item{plotlist}{plotlist}

\item{cols}{ncol}

\item{layout}{layout of plots.}
}
\description{
code from internet
}
