% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_density}
\alias{plot_density}
\title{Transformed specificity density plot}
\usage{
plot_density(df, ymax = 12)
}
\arguments{
\item{df}{A data.frame contains specificity, which from \code{ts_psi},
\code{ts_expr}.}

\item{ymax}{numeric. maximum in ylim. Default 12.}
}
\description{
Plot density curves of 9 methods.
}
\examples{
\dontrun{
result <- ts_psi(tmp_psi,
                 tissues = c("sample_A", "sample_B", "sample_C",
                             "sample_D", "sample_E", "sample_F",
                             "sample_G", "sample_H", "sample_I",
                             "sample_J", "sample_K", "sample_L",
                             "sample_M", "sample_N", "sample_O",
                             "sample_P", "sample_Q"),
                             identifier = "AS_events")
plot_density(result)
}
}
