% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{psi_seq_rank}
\alias{psi_seq_rank}
\title{Ranking psi according to equal-width-intervals}
\usage{
psi_seq_rank(df, n = 10, min = 0, max = 100)
}
\arguments{
\item{df}{data.frame.}

\item{n}{number of medium intervals to generated (all \code{n}+2). Default
10.}

\item{min}{psi under \code{min} will be set to 0.}

\item{max}{psi over \code{max} will be ranked highest.}
}
\value{
data.frame contain ranks from \code{df}.
}
\description{
Function to cut data.frame range into n equal-width interval points, maximal
point and minimal point, then rank them for their value. This function is
design for processing data, in which detecting differantial expression rely
on diffenrence, like PSI(delta psi).
}
