% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_prep.R
\name{quant_norm}
\alias{quant_norm}
\title{quantile normalization}
\usage{
quant_norm(df)
}
\arguments{
\item{df}{data.frame.}
}
\description{
Quantile-normalize data.frame. All 0 are set to NA, to exclude them from
quantile-normalization, then 0 values (the one set to NA) are set back to
0 finally.
}
