% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_prep.R
\name{rep_diff}
\alias{rep_diff}
\title{diff expression of replicates}
\usage{
rep_diff(df, tissues)
}
\arguments{
\item{df}{data.frame.}

\item{tissues}{charactors. Unique word-start identifiers to recognize
sample replicates.}
}
\description{
Function requires data frame with numeric. Difference between
replicates are calculated. \code{tissues} must be the unique word-start
identifier to recognize sample replicates. \code{identifier} is the colname
of gene names or AS events.
}
