% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_prep.R
\name{sort_dat_de}
\alias{sort_dat_de}
\title{Sort data.frame (decrease)}
\usage{
sort_dat_de(dat)
}
\arguments{
\item{dat}{data.frame, numeric.}
}
\description{
Sort rows of dataframe according row maximal value coordinate.
}
