% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_prep.R
\name{sort_dat_in}
\alias{sort_dat_in}
\title{Sort data.frame (increase)}
\usage{
sort_dat_in(dat)
}
\arguments{
\item{dat}{data.frame, numeric.}
}
\description{
Sort rows of dataframe according row minimal value coordinate.
}
