% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_Ee}
\alias{ts_Ee}
\title{Ee}
\usage{
ts_Ee(x)
}
\arguments{
\item{x}{data.frame.}
}
\description{
Function require a data frame with expression data, and give back a vector
with EEi values for each gene. If expression for one tissue is not known,
gene specificity for this gene is NA. Transformation: x/maxX.
}
\examples{
\dontrun{
ts_Ee(tmp_tpm)
}
}
