% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ts_expr}
\alias{ts_expr}
\title{Calculate specificity for a given gene expression data.frame}
\usage{
ts_expr(df, binary = "seq", n = 12, min = 0, max = 16, step = 1,
  trans = "log2_QN", tissues, identifier, na.del = TRUE, cutoff = 1,
  mingap = 2)
}
\arguments{
\item{df}{data.frame, which contain gene expression vaules.}

\item{binary}{"seq" or "quant". Binary method, "seq" refer to
equal-width-intervals and "quant" refer to equal-density-intervals.}

\item{n}{Integer. Cut the gene expression values into (n+2)
equal-density-intervals. Default 10.}

\item{min}{Numeric. The values under min will be graded into rank 0.
Defaut 0. Be careful when used with \code{trans}.}

\item{max}{Numeric. The values over max will be graded into rank (n+1).
Default 16. Be careful when used with \code{trans}.}

\item{step}{Numeric. Width of intervals in "seq" method. Be careful when
used with \code{trans}.}

\item{trans}{Charactor. "log2", "log2_QN", "QN". "QN" means
\code{quantile.normolize}.}

\item{tissues}{Vector of charactors, at leat 2. Analysed Tissues' unique
identifiers, and must keep away from "Tau", "Gini", "Tsi", "Counts", "Ee",
"Pem", "Hg", "Z", "Spm", "Mean" and "Max".}

\item{identifier}{Charactor. The colname of unique identifiers for row
symbols, like "gene_id" or "AS_events".}

\item{na.del}{Logical. Weather NA will be dropped. Default TRUE.}

\item{cutoff}{Numeric. Values under cutoff will set to 0(unexpressed) in
Specificity method Counts.}

\item{mingap}{integer. Minimal gap to generate binary pattern. Default 2.}
}
\value{
Specificity of gene expression values in \code{df}. A list with 2
data.frames, which contain raw values and binary pattern values and their
specificty values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee", "Pem",
"Hg", "Z", "Spm".
}
\description{
Function to detect tissue specific gene, and return a list with 2
data.frames, which contain raw values and binary pattern values and their
specificty values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee", "Pem",
"Hg", "Z", "Spm".
Rows with NA will be dropped. Of the binary pattern, all values betwwen min
and max in the data.frame will be graded into n equal-width-intervals or n
equal-density-intervals and assign the rank 0 to (n+1), respectively.
}
\examples{
ts_psi(tmp_tpm,
       tissues = c("sample_A", "sample_B", "sample_C", "sample_D",
                   "sample_E", "sample_F", "sample_G", "sample_H",
                   "sample_I", "sample_J", "sample_K", "sample_L",
                   "sample_M", "sample_N", "sample_O", "sample_P"),
                   identifier = "gene_id")
}
