% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ts_gene}
\alias{ts_gene}
\title{Get a tissue-specific gene subset by Ib and subtissues.}
\usage{
ts_gene(lst, Ib = 1, subtissues)
}
\arguments{
\item{lst}{List of 3 data.frame. one of them contains raw values(or psi values
with their specificty values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee",
"Pem", "Hg", "Z", "Spm"(named "raw"), the second contains rank values and binary
index(named "rank"), the third contains binary pattern values and index binary "Ib"
(named "bin"), which were generated from \code{ts_psi} or \code{ts_expr}.}

\item{Ib}{Integer. Cutoff of binary index, work with \code{dat_type}="rank".}

\item{subtissues}{a vector. The elements including in colnames of data.frame in lst.
The length of vector <= Ib.}
}
\value{
a list with 2 data.frames. one of them contains raw values(or psi values
with their specificty values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee",
"Pem", "Hg", "Z", "Spm"(named "rawgene"), the second contains rank values, binary
index and Type(named "rankgene").
}
\description{
Function to get a gene subset by Ib and subtissues in which genes are overexpressed,
and return a list with 2 data.frames, which contain raw values(or psi), rank values,
binary index, binary pattern values and their specificty values of 9 methods: "Tau",
"Gini", "Tsi", "Counts", "Ee", "Pem", "Hg", "Z", "Spm".
}
\examples{
res<-ts_expr(demo_tpm, n = 11,
       tissues = c("sample_A", "sample_B", "sample_C", "sample_D",
                   "sample_E", "sample_F", "sample_G", "sample_H",
                   "sample_I", "sample_J", "sample_K", "sample_L",
                   "sample_M", "sample_N", "sample_O", "sample_P"),
                   identifier = "gene_id")
ts_gene(res, Ib=2, subtissues=c("sample_A", "sample_B"))
}
