% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_Gini}
\alias{ts_Gini}
\title{Gini}
\usage{
ts_Gini(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Function require a vector with expression of one gene in different tissues.
If expression for one tissue is not known, gene specificity for this gene is
NA. Transformation: x*(N/(N-1)).
}
\examples{
\dontrun{
ts_Gini(c(1,2,3,2,8))
}
}
