% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_Hg}
\alias{ts_Hg}
\title{Hg, entropy}
\usage{
ts_Hg(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Function require a vector with expression of one gene in different tissues.
If expression for one tissue is not known, gene specificity for this gene is
NA. Transformation: 1-x/log2N.
}
\examples{
\dontrun{
ts_Hg(c(0,1,2,3,2,8))
}
}
