% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_index}
\alias{ts_index}
\title{Calculate specificity for a given numeric data.frame}
\usage{
ts_index(df, cutoff = 1)
}
\arguments{
\item{df}{data.frame of numeric.}

\item{cutoff}{Numeric. Values under cutoff will set to 0(unexpressed) in
Specificity method Counts.}
}
\value{
data.frame with original value with specificity index of 9 methhods.
}
\description{
Function to calculate specificity for a given numeric data.frame and return
a data.frame with original value with specificity index of 9 methhods.
}
