% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{ts_pattern}
\alias{ts_pattern}
\title{Calculate binary pattern for a given ranks data.frame}
\usage{
ts_pattern(df, mingap = 5)
}
\arguments{
\item{df}{data.frame of ranks.}

\item{mingap}{integer. Minimal gap to generate binary pattern. Default 3.}
}
\value{
data.frame with ranks with binary pattern and phenotype("DE" or "UC").
}
\description{
Function to calculate binary pattern(0/1) for a given ranks data.frame and
return a data.frame with ranks with binary pattern and phenotype("DE" or "UC").
}
