% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_Pem}
\alias{ts_Pem}
\title{Pem}
\usage{
ts_Pem(x)
}
\arguments{
\item{x}{data.frame.}
}
\description{
Function require a data frame with expression data, and give back a vector
with PEM scores. If expression for one tissue is not known, gene specificity
for this gene is NA. Transformation: x/(maxX).
}
\examples{
\dontrun{
ts_Ee(tmp_tpm)
}
}
