% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ts_psi}
\alias{ts_psi}
\title{Calculate specificity of psi}
\usage{
ts_psi(df, n = 20, min = 0, max = 100, tissues, identifier,
  na.del = TRUE, mingap = 5)
}
\arguments{
\item{df}{data.frame, which contain psi vaules (0-100). One column is the names of
symbols, like AS events id, etc.}

\item{n}{Integer. \code{n+2} equal-width-intervals be generated of all psi
values. Default 10.}

\item{min}{Numeric. The values under \code{min} will be graded into rank
0(unexpressed). Defaut 0.}

\item{max}{Numeric. The values over \code{max} will be graded into rank
\code{n+1}(highest). Default 100.}

\item{tissues}{Vector of charactors, at leat length of 2. Analysed
Tissues' unique identifier, and must keep away from "Tau", "Gini", "Tsi",
"Counts", "Ee", "Pem", "Hg", "Z", "Spm", "Ib", "Type", "Mean" and "Max",
exactly.}

\item{identifier}{Charactor, length of 1. The colname of unique identifier
for row symbols, like "gene_id", "AS_events", etc.}

\item{na.del}{Logical. Whether NAs will be dropped. Default TRUE.}

\item{mingap}{Integer. Minimal gap of generating binary pattern, Please
refer the paper. Default 5.}
}
\value{
list with 5 data.frames.
1) "raw" contain mean values of replicates (PSI and their specificty values
of 9 methods ("Tau", "Gini", "Tsi", "Counts", "Ee", "Pem", "Hg", "Z", "Spm"),
and Means of expression, Max values of expression, Ranges of expression,
and sd of expression.
2) "rank" contain rank pattern values, Binary Index, and type defined with
mingap.
3) "bin" contain binary pattern values, Binary Index, and type defined with
mingap.
4) "diff" contain maximal difference of replicates for each cell type, maximal
difference of all cell type, and sd of all difference.
5) "origin" contain raw data with replicates before any manipulation.
}
\description{
Function to find tissue-specific AS-events for a given data.frame.
}
\details{
Function to detect tissue specific gene, and return a list with 5
data.frames.
1) "raw" contain mean values of replicates (PSI and their specificty values
of 9 methods ("Tau", "Gini", "Tsi", "Counts", "Ee", "Pem", "Hg", "Z", "Spm"),
and Means of expression, Max values of expression, Ranges of expression,
and sd of expression.
2) "rank" contain rank pattern values, Binary Index, and type defined with
mingap.
3) "bin" contain binary pattern values, Binary Index, and type defined with
mingap.
4) "diff" contain maximal difference of replicates for each cell type, maximal
difference of all cell type, and sd of all difference.
5) "origin" contain raw data with replicates before any manipulation.
}
\examples{
res1 <- ts_psi(demo_psi, n = 20, min = 0.5,
               tissues = c("sample_A", "sample_B", "sample_C", "sample_D",
                           "sample_E", "sample_F", "sample_G", "sample_H",
                           "sample_I", "sample_J", "sample_K", "sample_L",
                           "sample_M", "sample_N", "sample_O", "sample_P",
                           "sample_Q"),
                           identifier = "AS_events")
res2 <- ts_psi(demo_psi, n = 10, min = 1,
               tissues = c("sample_A", "sample_B", "sample_C", "sample_D",
                           "sample_E", "sample_F", "sample_G", "sample_H",
                           "sample_I"),
                           identifier = "AS_events")
}
