% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_Spm}
\alias{ts_Spm}
\title{SPM score from TISGED}
\usage{
ts_Spm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Function require a vector with expression of one gene in different tissues.
If expression for one tissue is not known, gene specificity for this gene is
NA. Transformation: maxX.
}
\examples{
\dontrun{
ts_Spm(c(0,1,2,3,2,8))
}
}
