% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spe_meth.R
\name{ts_Tsi}
\alias{ts_Tsi}
\title{Tsi}
\usage{
ts_Tsi(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Function require a vector with expression of one gene in different tissues.
If expression for one tissue is not known, gene specificity for this gene is
NA.
}
\examples{
\dontrun{
ts_Tsi(c(1,2,3,2,8))
}
}
