% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_pattern.R
\name{bin_index}
\alias{bin_index}
\title{binary index}
\usage{
bin_index(x, mingap = 3)
}
\arguments{
\item{x}{Integer vector, Ranks.}

\item{mingap}{Minimal gap to generate binary pattern.}
}
\value{
binary index with "DE"(differential exprresion) or "UC"(unclear).
}
\description{
Function to generate binary index(the sum of \code{binary pattern}) of
ranked data. Define gaps as the diferrence of sorted vector(ranks,
low to high), and ranks over the maximal gap in the sorted vector set to 1,
otherwise set to 0, always select the first maximal gap.
If specify \code{mingap}, then ranks over the \code{mingap} set to 1, othewise
set to 0. And if there has no \code{mingap} in the vector, binary index set
to NA. Ranks value from function:\code{psi_seq_rank}, \code{expr_seq_rank},
\code{expr_quant_rank}. Ib is the sum of binary pattern if vector has mingap,
otherwise Ib = 0.
}
