% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_prep.R
\name{expr_max}
\alias{expr_max}
\title{maximal value of replicates}
\usage{
expr_max(x)
}
\arguments{
\item{x}{numeric.}
}
\description{
Function require a vector with numeric in different conditions.
Max is calculated taking in account tissues with 0 expression: 2+0+4=4.
}
