% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_density}
\alias{plot_density}
\title{Density plot of transformed specificity}
\usage{
plot_density(lst, ymax = 12)
}
\arguments{
\item{lst}{List of 3 data.frame. one of them contains psi values with their
specificty values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee", "Pem",
"Hg", "Z", "Spm"(named "raw"), the second contains rank values and binary
index(named "rank"), the third contains binary pattern values and
index binary "Ib"(named "bin"), which were generated from \code{ts_psi} or
\code{ts_expr}.}

\item{ymax}{Numeric. maximum of y limitation. Default 12.}
}
\description{
Plot density curves of 9 methods. Transformation methods in this function,
refers to paper, please!
}
\examples{
\dontrun{
result <- ts_psi(demo_psi,
                 tissues = c("sample_A", "sample_B", "sample_C",
                             "sample_D", "sample_E", "sample_F",
                             "sample_G", "sample_H", "sample_I",
                             "sample_J", "sample_K", "sample_L",
                             "sample_M", "sample_N", "sample_O",
                             "sample_P", "sample_Q"),
                             identifier = "AS_events")
plot_density(result)
}
}
