% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{heatmap for subset specificity expression}
\usage{
plot_heatmap(df, dat_type = "raw", specificity = c(0.8, 1),
  ts_method = "Tau", Ib = 1, ...)
}
\arguments{
\item{df}{List of 2 data.frame. Results of \code{ts_expr} or \code{ts_psi}.}

\item{dat_type}{"raw" or "bin". Plot raw data(psi/FPKM/RPKM/TPM) or
bin(binary index). Default "raw". "raw" work with \code{specificity}, and
"bin" work with \code{Ib}.}

\item{specificity}{numeric. Cutoff of specificity, work with
\code{dat_type}="raw". Default 0.8.}

\item{ts_method}{specificity methods. "Tau", "Gini", "TSI", "Counts", "EE",
"Hg", "Zscore", "SPM", "PEM".}

\item{Ib}{integer. Cutoff of binary index, work with \code{dat_type}="bin".}

\item{...}{parameters of pheatmap pkg.}
}
\description{
Wrapper function derived from pheatmap. Use the results of \code{ts_expr}
or \code{ts_psi} as Input.
}
\examples{
result <- ts_psi(tmp_psi,
                 tissues = c("sample_A", "sample_B", "sample_C",
                             "sample_D", "sample_E", "sample_F",
                             "sample_G", "sample_H", "sample_I",
                             "sample_J", "sample_K", "sample_L",
                             "sample_M", "sample_N", "sample_O",
                             "sample_P", "sample_Q"),
                             identifier = "AS_events")
plot_heatmap(result)
plot_heatmap(result, dat_type = "bin", Ib = 2)
}
