% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissSpe.R
\docType{package}
\name{TissSpe}
\alias{TissSpe}
\alias{TissSpe-package}
\title{TissSpe: A package for calculating specificity.}
\description{
The \code{TissSpe} package provides 4 functions: \code{ts_psi},
\code{ts_expr}, \code{plot_density} and \code{plot_heatmap}.
}
\section{TissSpe functions}{

\code{ts_psi}: caculate specificity of psi.
\code{ts_expr}: caculate specificity of gene expression.
\code{plot_density}: plot density of results from \code{ts_psi} or \code{ts_expr}.
\code{plot_heatmap}: heatmap of sub-results from \code{ts_psi} or \code{ts_expr}.
}

