% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ts_psi}
\alias{ts_psi}
\title{Calculate specificity of psi}
\usage{
ts_psi(df, n = 10, min = 0, max = 100, tissues, identifier,
  na.del = TRUE, cutoff = 1, mingap = 3)
}
\arguments{
\item{df}{data.frame, which contain psi vaules.}

\item{n}{Integer. Cut the psi values into (n+2) equal-width-intervals.
Default 10.}

\item{min}{Numeric. The values under min will be graded into rank 0.
Defaut 0.}

\item{max}{Numeric. The values over max will be graded into rank (n+1).
Default 100.}

\item{tissues}{Vector of charactors, at leat 2. Analysed Tissues' unique
identifier, and must keep away from "Tau", "Gini", "Tsi", "Counts", "Ee",
"Pem", "Hg", "Z", "Spm", "Mean" and "Max".}

\item{identifier}{Charactor. The colname of unique identifier for row
symbols, like "gene_id" or "AS_events".}

\item{na.del}{Logical. Weather NA will be dropped. Default TRUE.}

\item{cutoff}{Numeric. Values under cutoff will set to 0(unexpressed) in
Specificity method Counts.}

\item{mingap}{integer. Minimal gap to generate binary pattern. Default 3.}
}
\value{
Specificity of psi values in \code{df}. A list with 2 data.frames,
which contain raw values and binary pattern values and their specificty
values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee", "Pem", "Hg",
"Z", "Spm".
}
\description{
\code{ts_psi} returns the list of original data with specificities of 10
methods.
}
\details{
Function to detect tissue-specific AS events, and return a list with 2
data.frames, which contain raw values and binary pattern values and their
specificty values of 9 methods: "Tau", "Gini", "Tsi", "Counts", "Ee", "Pem",
"Hg", "Z", "Spm". Rows with NA will be dropped. Of the binary pattern, all
values betwwen min and max in the data.frame will be graded into n
equal-width-intervals and assign the rank 0 to (n+1), respectively.
}
\examples{
ts_psi(tmp_psi,
       tissues = c("sample_A", "sample_B", "sample_C", "sample_D",
                   "sample_E", "sample_F", "sample_G", "sample_H",
                   "sample_I", "sample_J", "sample_K", "sample_L",
                   "sample_M", "sample_N", "sample_O", "sample_P",
                   "sample_Q"),
                   identifier = "AS_events")
}
